Fix crash in JNI proxy cache (#378)
Updated JNI templates for proxy cache to fix a bug in the destructor of
CppProxyBase class. Previous implementation incorrectly omitted
"type_key" string from the cache key used in the destructor, thus
failing to erase the cache entry. That led to crashes in operator==()
later when a stale entry with a deleted Java object was accessed.

Updated smoke tests. A functional regression test seems to be
infeasible. Consistent reproduction of the bug relies on the Java object
referred from the stale cache entry to be garbage-collected, and GC
action in JVM is non-enforcible and thus not consistent enough to be
reproducible in tests.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>