Use try-finally for "fromFfi" conversions in Dart
Updated Dart templates to wrap "fromFfi" conversion calls in
`try-finally` construct, releasing the handle in the `finally` block.
This way the handle is released even if the conversion function throws.

No conversion function is expected to throw normally, but they can throw
if malformed data is passed from C++ code. The upcoming Locale type
implementation is especially sensitive to malformed data (but in general
this can happen to most supported types).

See: #372
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>