Fix disambiguation of method parameters in Dart
Updated LimePath class to expose "disambiguator" (user-invisible tag
that disambiguates method overloads) as public. Also changed `child()`
methods of LimePath to "inherit" the disambiguator when a child path is
created.

Updated DartNameResolver to use this new functionality to correctly
determine which method a parameter belongs to, in case of method
overloads (previously it was impossible due to disambiguation
information being lost in `LimeParameter.path`). This fixes doc comment
links to such parameters in Dart (as seen in the updated smoke test).

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>