Fix cross-package inheritance from interfaces in Java
Updated Java and JNI model builders and templates to make classes that
inherit from interfaces to implement those interfaces directly instead
of reusing the internal "Impl" classes. This fixes the compilation issue
that arises when the class and the interface are in different packages.

Added functional test. Updated smoke tests.

Resolves: #394
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>