Add @immutable attribute in Dart
Updated Dart template for structs to add `@immutable` attribute to
structs that are marked as `@Immutable` in IDL. Also added `const`
modifier to constructors of such structs in Dart.

Updated smoke tests. No new functional tests are needed as this change
is aimed for Dart static analysis tools and so are expected to have no
impact to Dart compilation and run-time behavior.

Resolves: #433
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>