Refactor qualifiers of methods and classes in Java model
Updated Java language model to represent "static" and "native"
qualifiers for methods as Boolean flags instead of a set of enum values.

Updated Java language model to represent "static" and "final"
qualifiers for classes as Boolean flags instead of a set of enum values.

Updated Java templates for both of these changes.

These changes allow for more fine-grained control over these qualifiers
on template level. In particular, it will be necessary to suppress
"native" and "final" qualifiers if the "stubs" mode is enabled in the
future "stubs" mode implementation.

See: #436
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>