Fix name clash for @Dart(Default) constructors (#474)
Updated DartNameResolver to give a distinct "$init" name to constructors
marked with `@Dart(Default)`. This does not affect any publically
visible API, but only the names of the private functions. This resolves
a name clash that was happening when such constructor had a same IDL
name as another constructor in the same type.

Added functional and smoke tests.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>