Fix compilation for cross-package interface inheritance in Java
Updated Java model to generate Impl classes for child intefaces that are completely independent from the Impl class of
the parent inteface. This solves a compilation issue when child and parent interfaces are in different packages (thus
making usage of package-private parent Impl class incorrect).

Added smoke and functional tests.

Resolves: #491
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>