Fix "hasAnyComment" predicate for C++ templates
Updated "hasAnyComment" predicate for C++ templates to check comments for not being empty through the C++-specific text
of these comments instead of through the whole comment (which also includes text for other platforms).

Added a smoke test where the C++ comment is blank while the whole comment is not.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>