Delay exception throwing in JNI until the end of method scope
Updated JNI template for functions throwing an exception to use
the new JniExceptionThrower class to delay throwing the
Java exception until the very end of the function scope.

This fixes the issue where a C++ proxy (for a Java interface
passed as a function parameter) tries to access JNI API but
there is a Java exception pending already.

Updated smoke tests. Added a functional test.

Resolves: #605
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>