Add workaround for Dart-C++-Dart object round trip (#609)
In the current implementation, sending two native Dart objects (lambdas or interface implementations) to C++ and then
back again could sometimes cause a crash when the first object is destroyed in C++ but its Dart deleter is delayed by
the callback queue.

Updated related Dart templates to perform `instanceOf` checks before casting objects obtained from the reverse cache.
This properly treats the scenario as cache miss.

This is a temporary workaround. The full fix will be done in the scope of #606 later. As the issue does not arise in
practical scenarios, the full fix has a low priority.

Also restored Dart builds, without AddressSanitizer.

See: #606
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>