Add tests for potential Dart name clash
Added smoke and function tests for an interface with a "dispatch()" function in Dart. This function name has a potential
for a name clash with a hard-coded function with the same name in the Dart C++ proxy. However, the signatures are
different enough and the name clash does not happen.

Resolves: #628
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>