Generate separate "Export.h" file for "common" sources
Added new command line parameter "cppexportcommon" to specify a
separate export macro name for "common" sources in C++.

Updated C++ generator to create a separate "ExportCommon.h"
header file for this separate "common" export macro. This
separate file is created and used even if the new command line
parameter is not specified and the macro name is the same
as the one for "main" sources.

See: #650
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>