Move Dart "reverse cache" from Dart code to Dart FFI code (#667)
Dart generated bindings contains a mechanism for maintaining referential equality (i.e. guaranteeing that the same C++
object sent to Dart side twice produces the same Dart object). This mechanism is a "reverse" cache with C++ raw pointer
as a key and Dart object (represented by an isolate-unique "token").

This mechanism mostly works correctly. However, due to threaded nature of C++-to-Dart callbacks, removal of entries from
this cache might be delayed if initiated from C++ side. This delay, in turn, can lead to false cache hits, when the C++
memory address of the old object is reused.

This commit moves this mechanism to Dart FFI generatedcode. This way there is no delay on cleanup and false cache hits
are eliminated.

Resolves: #606
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>