Compute enumerator implicit values dynamically in LIME model
Updated LimeEnumerator to keep track of its preceding sibling inside to the same enumeration and to use this sibling to
dynamically compute an implicit value when necessary.

Previous implementation pre-computed all enumerator implicit values at the moment of model loading. However, this
approach works incorrectly in Java and Dart if some of the enumerators are skipped.

See: #703
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>