Generalize skip-tag model prefiltering
Moved filtering of LimeModel by skip-tags out of individual generators and into a single place in Gluecodium.kt, as it
is the common step for all generators (except for the special case of LIME generator).

Moved LimeModelFilter class to :lime-runtime subproject. This will allow to use internal tools of :lime-runtime for
filtering in upcoming implementations.

Refactored LimeModelFilter to filter reference map in addition to the element tree that was already filtered before.

See: #715
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>