Change Dart instance cache to use weak references
Implementing automatic finalizers in Dart requires that all caches in the generated code that hold a Dart object should
hold it be a weak reference (holding it by a strong reference prevents the finalization).

For the "token cache" that uses Dart objects as keys the `Expando` class fulfills the weak reference requirement.

For the instance cache itself, there is no way in pure Dart to hold a (non-opaque) weak reference to a Dart object.
However, there is such functionality in FFI APIs (`Dart_NewWeakPersistentHandle_DL()` and related functions). Therefore
this cache was moved to FFI side.

Smoke tests are updated in separate commit.

See: #890
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>