Fix "moveToCType" usage in Swift
Updated Swift templates to avoid usage of `moveToCType().ref`
antipattern. The result of `moveToCType()` call needs to
be stored in a local varable instead, to extend its lifetime to
the lifetime of the current scope and thus preventing
the releasing of `.ref` handle long enough for it to be used.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>