Add "positional defaults" for Java structs (#801)
Updated JavaStruct.mustache to add support for `@Java(PositionalDefaults)` attribute. Similarly to what this attribute
already does for Dart, the generated code with this attributes imitates constructors with parameter default values in
Java. Given the limitations of Java language, this means generating as many separate constructors as there are
initialized fields.

Added smoke and functiona tests.

Resolves: #795
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>