Add `@Optimized` attribute to LIME IDL grammar (#863)
Updated LIME IDL grammar definition and loader to support `@Optimized` attribute. This attribute is intended for use on
type references to `List<>` type, indicating that bindings with on-demand fetching of list elements need to be
generated.

Initially this attribute is only going to be supported for:
* function return types
* read-only properties
* fields of immutable structs

Added LimeOptimizedListsValidator to ensure these usage constraints. Added unit tests for the new validator.

Added support for this attribute to C++ generator. Added corresponding smoke test.

See: #837
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>