Introduce `NativeBase` in Dart  (#876)
* Introduce `NativeBase` in Dart

Added internal `NativeBase` class to generated code in Dart. This class serves the same purpose similarly named classes
in Java and Swift generated code already do: providing a single base class to obtain the opaque handle from.

This fixes the runtime issue where declaring an inteface and a child class inheriting from it in IDL, and then passing
an instance of that child class from C++ to Dart and back to C++, created a double-wrapped object on C++ side instead of
correctly resolving into the original C++ object.

Added functional tests that also serve as regression tests for that issue. Smoke tests are updated in a separate commit.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>

* Update smoke tests for Dart NativeBase addition

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>