Run Dart functional test on Ubuntu 20 on CI
Recent update to GitHub Actions CI image for Ubuntu 18 broke one of the functional tests for Dart. The same test does
not fail locally, does not fail on the old image (which is no longer accessible), and does not fail on Ubuntu 20.
Considering all of the above, this looks to be a CI configuration problem, not an issue in the generated code.

As a workaround, updated CI config to use Ubuntu 20 image for Dart functional tests CI jobs. Also updated CMake script
for these tests to autodetect "libasan" path, as it differs slightly between different versions of Ubuntu.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>