Fix usage of `Locale` hash function in C++ sets and maps (#941)
Inverted the internal condition in `CppLibraryIncludes.hasStdHash()` predicate so that it lists types for the `true`
result explicitly and returns false otherwise.

The list of `true` types for this predicate is not expected to change. However, each new "basic" type added to LIME
model has to return `false` from this predicate. The new inverted condition accomplishes this automatically. This also
fixes the predicate erroneously returning `true` for the `Locale` type.

Added/updated smoke and functional tests to cover the usages of `Locale` hash function in C++.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>