Fix CBridge warning about C99 feature
Updated "CBridgeFunctionDefinition" template to use explicit field name for both fields when initializing "return value
with error" struct. Using an explicit name for value/error field is required as it is a union. Mixing nameless and named
fields in initialization is a non-standard C99 feature, which produced a compiler warning. The updated template fixes
the warning.

Updated smoke tests.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>