Improve smoke tests to use snippets as reference
Updated SmokeTest and NiceErrorCollector classes to support comparing generated code against reference code snippets
(partial files) instead of full reference files as before. This approach is backwards-compatible with the old one, as
full files are still valid "snippets".

This allows excluding parts of the reference file that are not relevant to the feature-under-test in each test case.
Which, in turn, reduces maintenance load by reducing the number and the size of reference files affected by each change.

As a demonstration of this approach, updated "basic types" smoke test to use snippets.

See: #953
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>