Fix Dart FFI missing include for top-level lambdas
Updated Dart "FfiImplementation" template to have "FinalizerData.h" include when lambdas are present. This fixes a
compilation issue where this include is missing if there are no classes or interfaces present, only lambdas.

The issue was introduced with "automatic finalizers for lmabdas" feature, which is not released yet, so no separate
changelog entry is added for the fix itself.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>