Add support for specifying C++ type for `Date` type
Added support for `@Cpp(Type)` attribute that can be applied to any type reference of `Date` type in IDL to specify an
`std::chrono::timepoint<>` type to be used. If this attribute is not added, `std::chrono::system_clock::timepoint` is
still used as before. If this attribute is added, an alternative `timepoint<>` type is used in generated C++ code with
no visible changes to public API in other output languages.

Added smoke and functional tests. The tests are based on the proposed use case of using monotonic clock time point
`std::chrono::steady_clock::timepoint`.

Resolves: #965
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>