Fix the crash for overloads of a skipped function
Updated LimeModelFilter to restore "ambiguous" keys in the reference map if some of the related non-ambigous keys are
still present in the map after filtering. This fixes the issue where Gluecodium crashes while trying to parse a
documentation comment on a function parameter when another overload of this function is skipped.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>