Move tag-filtering to generators
Updated the filtering of the LIME model to be done in one step (in the generators), instead of two steps (also in
Gluecodium.kt class) as before. The "old" approach separated tag-filtering and platform-filtering, the new approach
combines them in one place (predicates extracted to LimeModelSkipPredicates). This is needed for implementing additional
logic that can depend on both types of filtering at the same time.

See: #980
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>