Fix overloads validation and related issues
Fixed the way LimeOverloadsValidator is called in generators. The "old" way was
not actually validating any functions. Fixed the following issues revealed by
the restored validation.

Updated LimeSignatureResolver to take nullability of function parameters when
creating a signature for a function. This fixes the issue where functions
differing by nullability only were incorrectly treated as overloads (they are
still correctly overloads in Java). Updated name mangling in CBridge
and FFI name resolvers to properly handle signatures with nullabililty.

Removed validation of constructor overloads, because it was incorrect. It should
vary by platform, and needs its own validator(s). Issue #1034 was created to
address that.

Smoke tests are updated in a separate commit.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>