Restore validation of constructor overloads
Updated LimeSignatureResolver to detect overloads correctly:
* For function overloads, only container's own functions are taken into account,
not parent container's functions (in case of inheritance).
* For constructor overloads, the same applies, except function names are ignored
(as all constructors have the same "name").
* For Swift constructor overrides detection, parent container's functions are
taken into account. Moved the dedicated predicate to SwiftSignatureResolver as
it is not needed elsewhere.

Updated LimeOverloadsValidator to optionally validate constructor overloads
(needed in Java and Swift, not needed in C++).

Updated LimeOverloadsValidatorTest unittests accordingly.

Resolves: #1034
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>