Fix overloads validation in Dart
Fixed LimeModelFilter to properly update the reference map with the newly
created elements, in addition to removing the skipped elements as before. This
fixes a bug in Dart overloads validation where skipped functions were
incorrectly taken into account when calculating overloads.

Added a new smoke test as a regression test.