Make implicit all-fields constructor private in Dart
Extracted Dart generator predicates into a stand-alone file, similarly to what
is already done for predicates in other generators. Added a predicate to check
whether the implicit all-fields constructor of a struct needs to be "private".

Updated Dart templates to make use of this new predicate.

This fixes an issues where a mutable struct with "field constructors" defined
had a public all-fields constructor, even though such constructor is not
required (it is still required and public for `@Immutable` structs).

Added smoke and functional tests.

Resolves: #1090
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>