Fix the loss of ref equality for interface inheritance in Dart
Updated Dart `toFfi()` conversion functions functions to try downcasting from
the parent interface type to the child interface types when creating a proxy,
based on the results of "descendant interfaces" analysis.

Updated DartDeclarationImportResolver to add imports needed for the new code,
but only when necessary.

Added smoke and functional tests.

Resolves: #1134
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>