Fix compilation issue for nested structs in Dart
Updated Dart struct templates to explicitly refer to `allfieldsConstructor`
value with '.' dot notation, instead of trying to fetch it from the current
context.

The way Trimou template engine works, `allfieldsConstructor` being `null` in the
inner of two nested structs results in a fall-through to the same value of the
outer nested struct. Fortunately, this behavior does not apply to empty
collections, only to `null` values, so most parts of the templates are safe with
regard to nesting.

Added smoke and functional tests as regression tests.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>