Detect parameterless field constructors in Dart name resolver
Updated DartNameResolver to detect a parameterless field constructor when
resolving names for field default values of struct type. Previously it was
hard-coded to always use `withDefaults()` constructor name.

This fixes a compilation issue when using a struct with `field constructor()` to
initialize a field of another struct.

Updated existing smoke and functional tests to serve as regression tests for
this issue.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>