Make Java "automatic" constructors forward-compatible
As a part of the effort to make automatically-generated struct constructors
forward-compatible, updated Java predicates and templates to omit the all-fields
constructor for mutable struct.

This way only a constructor for uninitialized fields is generated, making field
addition or removal a non-breaking change (as long as the added/removed field
has a default value).

This change does NOT affect:
* `@Immutable` structs
* structs with `@Java(PositionalDefaults)
* structs with explicitly defined constructors

Updated functional test IDL files with explicit constructrs to preserve APIs
that are used in tests.

Smoke test files are updated in a separate commit.

See: #1178
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>