Make Dart "automatic" constructors forward-compatible
As a part of the effort to make automatically-generated struct constructors
forward-compatible, updated Dart predicates and templates to make the all-fields
constructor private for mutable structs.

This way only a constructor for uninitialized fields is generated, making field
addition or removal a non-breaking change (as long as the added/removed field
has a default value).

As a consequence, the "initialized-fields" constructor for mutable structs is
now nameless, instead of `withDefaults()` as before.

This change does NOT affect:
* `@Immutable` structs
* structs with `@Dart(PositionalDefaults)`
* structs with explicitly defined constructors

Updated Dart functional tests to use a nameless constructor instead of old
`withDefaults()` where appropriate. Updated functional test IDL files with
explicit constructors to preserve APIs that are used in Dart tests.

See: #1178
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>