Fix redundant imports in Dart for some external types
Updated DartImportResolver to omit imports of type conversion functions for
types which have an `external.dart` descriptor and are used as a type parameter
for a collection type (list, set, or map) instead of being used directly.

For such usage the collection conversion functions are enough, and the functions
for the type itself are not needed.

Updated smoke tests to cover this scenario.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>