Validate internal fields in "strict" mode (#1213)
Updated LimeStructsValidator with an additional predicate for the "strict"
validation mode for struct. If a struct has an internal field that does not have
a default value specified for it, this struct should have an explicit
constructor. This applies only for mutable structs, as `@Immutable` structs have
to have an explicit constructor anyway in the "stict" mode.

See: #1178
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>