Fix "internal" inheritance issue in Java
Updated Java templates to force methods inherited by a class from an interface
to be always public. This fixes a compilation issue when both the class and its
interface parent are `internal`. Java treats all interface methods as "public",
even if the interface itself is not, so overriding interface methods with
lower-visibility class methods is a compilation error.

Added smoke and functional tests.

Resolves: #1222
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>