Improve field deprecation logic in Swift
As a part of the effort to make automatically-generated struct constructors
forward-compatible, updated Swift predicates and templates to generate an
addition "non-deprecated-fields" constructor for structs with deprecated fieds.

Together with a Swift deprecation attribute on the all-fields constructor for
such structs, this helps to support a better deprecation path for struct fields.
This change does not affect structs with explicitly defined constructors

Added SwiftFieldsValidator to check that all deprecated fields have default
values defined. Added unit tests for the new validator.

Added/updated smoke tests accordingly.

Resolves: #1178
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>