Improve support for `@Platform(Skip)` attributes on struct fields
Updated JNI/CBridge/FFI templates for struct conversion in "platform->C++"
direction to use the field's default value if the field itself is skipped in the
platform code. This makes "platform skip" attributes available on fields for
most cases.

The only unsupported corner case is when the field does not have a default value
defined, and the struct has some immutable fields (including when the struct is
itself `@Immutable`). Updated LimeSkipValidator to produce a validation error
for this corner case. Added new unit tests for the validator accordingly.

Added new smoke and functional tests for the newly supported "platform skip" use
cases.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>