Add missing predicates for generic type conversions in Dart
Updated DartGenerator to correctly pass all generator predicates to the
TemplateEngine when generating code for generic type conversions. This fixes an
issue when two types with the same name are used in a collection type, but their
conversion functions are (incorrectly) not prefixed with aliases in
`generic_types__conversion.dart`, leading to a compilation error.

Added smoke and functional tests.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>