Add support for `@NoCache` attribute in Java (#1245)
Added experimental support for `@NoCache` attribute in Java for classes and
interfaces. A type with this attribute will avoid caching on any language
boundary transition. This breaks the referential equality invariant, but frees
up some resources.

Currently added only for Java, as a stop-gap solution for an internal issue. The
issue is that whena lot of small objects are sent from C++ to Java, it can lead
to an overflow of JNI weak reference table. This happens on some older Android
devices where either the table is confiugured to be small, or the hardware is
too slow to fire the GC process promptly.

As these small objects are not intended to be compared for referential equality,
this `@NoCache` approach should solve the issue. The full implementation for
other languages, as well as more tests, will follow in a later commit.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>