Don't generate export macro for nested constants in C++
Updated C++ Mustache templates to skip `EXPORT` macro for constants that are
nested inside a class/interface/struct. The outer type has the same macro anyway
and this is enough to ensure symbol visibility for the constant.

Only stand-alone constants (declared inside `types` in IDL) are getting the
`EXPORT` macro now. This fixes a compilation issue for nested constants on
Windows platform.

Updated affected smoke tests.

Resolves: #1244
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>