Convert `Date` literals at generation instead of at parsing
LIME `Date` literals are indistiguishable from `String` literals at parsing
time. The initial implementation tried to use the type of the field of the
constant to make this distinction. However, this approach does not work for type
aliases, as they cannot be resolved until the whole model is parsed.

Due to this, the "old" approach generated uncompilable code with type aliases.
The "new" approach is to:
* treat all string literals the same at parsing time
* during validation, check string literals assigned to `Date` type field/const
to be ISO 8601 compliant
* do the actual `Date` literal conversion in the name/value resolvers

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>