Use own CBridge functions in a child class in Swift (#1314)
Updated Swift and CBridge templates to generate dedicated bindings for functions
in a class inherited from an interface. Previous implementation reused bindings
of the parent inteface, which let to compilation errors if that interface was
placed in a separate Swift framework.

Added smoke tests. Existing "inheritance" functional tests are covering this
scenario already and are passing.

See: #1312
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>