Use own FFI functions in a child class in Dart (#1315)
Updated Dart and FFI templates to generate dedicated bindings for functions
in a class inherited from an interface. Previous implementation reused bindings
of the parent inteface, which let to compilation errors if that interface was
placed in a separate Dart library.

Added smoke tests. Existing "inheritance" functional tests are covering this
scenario already and are passing.

Resolves: #1312
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>