Disambiguate LIME model paths automatically
Updated `AntlrLimeModelBuilder` and `AntlrLimeReferenceResolver` to
detect ambiguous paths automatically and add a disambiguation suffix for
them. Previous implementation only added the suffix for functions and
did so regardless of the path being ambiguous or not.

This enables correct preservation of ambiguous paths when filtering the
reference map in `LimeModelFilter`.

Added a regression smoke test that fails with the "old" approach, as it
incorrectly filters out too many reference map entries.

Resolves: #1313
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>