Add `LimePath.toAmbiguousString()` function
Added `toAmbiguousString()` function to `LimePath` class. The function creates a
string representation of the path where the disambiguation suffix is omitted.

Updated resolution logic for reference links in documentation comments to use
this new functions instead of `LimeNamedElement.fullName` in most places. For
the current implementation, it is just a refactoring, with no change of
functionality.

However, this refactoring decouples the link resolution logic from the
assumption that most paths don't have a suffix. With the assuption removed, a
new implementation can be done that introduces disambiguation suffixes for any
element, not just the functions as before.

See: #1313
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>