Add preliminary support for `@Internal` attribute (#1348)
Added support for `@Internal` attributes. Currently this attribute has
the same effect as `internal` visibility, with a limitation that the
attribute only work on type declarations (i.e. it doesn't work on
functions, properties, fields, and constants).

The attribute could be used interchangeably instead of- or redundantly
together with `internal` visibility. This is intended to help a gradual
phase-out of the visibility syntax.

See: #1333
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>