Make "internal" fields private in Dart
Updated Dart name resolver to resolve "internal" visibility to "_". This makes
the affected elements file-private in Dart. The only remaining elements that are
affected by this are struct fields.

This prevents accidental leaks of internal APIs to public (the leaks happen due
to lack of real "internal" visibility in Dart). Making the fields private
instead of just skipping them is necessary to preserve the data in the
non-public field for language boundary transitions.

See: #1333
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>