Fix compilation for "overloading over inheritance"
Fixed varios compilation issues when "overloading over inheritance". For
C++, adding a `using` directive for the parent functions in the child
class. For JNI and CBridge, adding a signature suffix to the C
functions.

For Dart FFI no changes, since it has this logic in place already, due
to overloading being absent from the Dart language as a concept
entirely.

Resolves: #1401
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>