Reduce visibility propagation (#1457)
Removed the logic for propagating visibility from outer to inner
elements from the AntlrLimeModel builder. For most languages, marking
the inner elements "internal" is redundant if the outer element is
already marked.

Added visibility propagation on case-by-case basis where needed:
* in Dart for nested types, as there is no nesting in Dart
* in Swift for types nested inside interfaces, as there is no nesting
inside Swift protocols

For both cases, the propagation affects types only. Non-type elements
are not affected.

This both simplifies the visiblity logic, and ensures that "global" and
"per-platform" visibilities are handled in the same way.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>