HERESUP-3274: Stop callback queue on main isolate exit
When callback is scheduled after main isolate is finished then application will deadlock on wait for callback to be accepted.

If callback queue is destroyed then scheduling callback fails and does not block scheduling thread.

Connect lifetime of the callback queue with the lifetime of isolate which executes those callbacks.

Signed-off-by: Mykhailo Diachenko <ext-mykhailo.z.diachenko@here.com>