Ensure that conversions are defined for lambdas defined in struct
In the case of Swift language, when lambda was defined in
a structure, then the code could not compile because of
missing conversion functions. Interestingly, that was not
the case for classes.

After investigation it turned out, that the root cause is
related to missing inclusion of SwiftLambdaConversion
template in SwiftStructConversion.

This change:
 - adds missing inclusion of lambda conversions in the
   mentioned template (SwiftStructConversion.mustache)
 - implements functional tests for Swift to check that
   the fix works as expected
 - implements functional tests for Dart and Android to
   also cover the tested case in these technologies
 - updates the reference files for smoke tests related
   to definition of nested fields in structs

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>