Extend smoke tests to show invalid behaviour of Dart generator
The behaviour of generator for Dart is invalid in the case
of types that are nested in another type that is marked as
internal.

Such nested types of an internal type shall not be exported.
In order to show the invalid behaviour the smoke tests were
added.

In next commit Dart generator will be fixed and then the exports
visible in 'smoke.dart' file will be removed.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>