Extend smoke tests for accessors C++ attribute
This change is used to extend smoke tests for
structures that use accessors C++ attribute.

The goal of this change is to show the invalid
behavior of C++ generator. The setters and getters
use by-value returns and parameters even for
non-trivial types.

In the next patch the generator will be adjusted.
The smoke test will be used to show the new behavior.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>