Dart: generate omitted documentation of properties
When users want to document a property
in LIME files, they can use structured
comments, which consist of the general
information about property and annotations
related to getters and setters.

In the case of Dart generator, the output
contained only comments related to getters
and setters. The general information about
property was not stored in the output files.

This change adjusts 'DartRedirectProperty'
mustache template to output also the general
information about property if present.

The behaviour in the case of Dart generator
has been aligned with C++ and Java generators.
The general information about property is added
to comments of getter and setter.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>