Docs: document Dart-specific behavior of internal tag (#1611)
The behavior of internal tag for Dart is different from
that for Java/Swift. In Dart, the types annotated as internal
do have internal visibility.

However, functions, properties, constructors and field constructors
are not generated. These elements are skipped like in the case of
skip attribute.

This change adds this information, because the discrepancy between
languages may be surprising for the users.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>