Extend smoke tests to show invalid behavior of attributes converter
In the case of internal attribute platforms can be specified
inside the annotation. However, when quotation marks are not
present for each listed platform, then the behavior of converter
is invalid.

This change adds omitted files for smoke tests.
Please note, that the visibility for mentioned types is as follows:
 - JavaInternalClassRev -- shall be internal only for Java, but
                           it is internal also in Swift and Dart.
 - SwiftInternalClassRev -- shall be internal only for Swift, but
                            it is internal also in Java and Dart.
 - DartInternalClassRev -- shall be internal only for Dart, but
                           it is internal also in Swift and Java.

In next patches the logic will be fixed. The mentioned files will
be used to show that the behavior has been adjusted.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>