Document a way to get consistent internal field constructor
The internal annotation works differently in the case of Dart.
When a field constructor is annotated as internal, then it is
not generated in Dart. That's not the case for Java/Swift. In
the mentioned langagues, the construtor becomes package-private
or internal.

In order to achieve the consistent behavior between code generated
for each of the languages, the user may use renaming of elements
in Dart to ensure, that the generated constructor is private.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>