cmake-tests: Fix test failing on iOS
The following test was failing:
 'dependent-frameworks-with-shared-modules-dependent-on-object-libs'

The test case was failing with preset for hidden
visibility enabled. The root cause of the failure
was related to the fact, that CppExportMacro works
only when appropriate compile definition is provided.

By default 'generate.cmake' provides the definition
only for shared libraries. The target from the test
was OBJECT. Therefore, 'scenario 2' from the documentation
of 'generate.cmake' was needed. It requires the user of the
function to manually query and set compile definitions.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>